pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-scheduler"
version = "1.0.0"
description = "Scheduled AI tasks for Claude Code"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "argcomplete>=3.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "rich>=13.0.0",
    "textual>=0.50.0",
    "gspread>=6.0.0",
    "google-auth>=2.28.0",
    "google-auth-oauthlib>=1.2.0",
    "authlib>=1.3.0",
    "itsdangerous>=2.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "gunicorn>=22.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "httpx>=0.27.0"]

[project.scripts]
cs = "claude_scheduler.cli:main"
claude-scheduler = "claude_scheduler.cli:main"
cs-mcp = "claude_scheduler.mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_scheduler"]
